+++ src/main/java/subway/application/dto/SectionAddDto.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.application.dto;
 
 public class SectionAddDto {
     private Long lineId;
