+++ src/main/java/subway/application/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.application.dto;
 
 public class StationRequest {
     private String name;
