+++ src/main/java/subway/domain/LineRepository.java
@@ -1,6 +1,7 @@
-package subway;
+package subway.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.domain.Line;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
 }
