+++ src/main/java/subway/domain/Section.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.domain;
 
 import javax.persistence.*;
 import java.util.Objects;
