+++ src/main/java/subway/domain/exceptions/CanNotAddSectionException.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.domain.exceptions;
 
 public class CanNotAddSectionException extends RuntimeException {
     public CanNotAddSectionException(String message) {
