+++ src/main/java/subway/ui/StationController.java
@@ -1,7 +1,10 @@
-package subway;
+package subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.application.dto.StationRequest;
+import subway.application.dto.StationResponse;
+import subway.application.StationService;
 
 import java.net.URI;
 import java.util.List;
