+++ src/main/java/subway/ui/SubwayAdviceController.java
@@ -1,8 +1,9 @@
-package subway;
+package subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import subway.domain.exceptions.CanNotAddSectionException;
 
 @RestControllerAdvice
 public class SubwayAdviceController {
